2006-02-06  Peter Dennis Bartok  <pbartok@novell.com>
	* XplatUIX11.cs: 
	  - The preceeding 64bit fixes had a bug: ChangeProperty expects the
	    number of bits for the property, not the number of bytes. The
	    change to provide IntPtr.Size broke 32bit. (64bit was also wrong
	    but would not crash since it specified 8 bits instead of 4 bits)
	  - More 64bit fixes: Switched all atoms from int to IntPtr (they are
	    defined as XID -> long in the C headers)
	  - Removed 'new IntPtr((int))' and '(IntPtr)' casts from all NetAtoms 
	    references since those are now IntPtr to begin with
	  - Switched all Atom.XXX 'int' casts to IntPtr casts
	  - Fixed XGrabPointer DllImport signature to work for 64bit (cursor arg)
	  - Fixed XInternAtom signature for 64bit, now returns an IntPtr
	  - Added XChangeActivePointerGrab DllImport (for X11DnD)
	* X11Structs.cs:
	  - Changed 'int' type for Atoms in XEvent structures to IntPtr
	  - Changed atom in HoverStruct to be IntPtr
	* X11DnD.cs:
	  - Removed local DllImports, switched code to use those from XplatUIX11
	  - Removed/fixed casts related to the switch of Atom to be a IntPtr


svn path=/trunk/mcs/; revision=56603

+++ mcs/class/Managed.Windows.Forms/System.Windows.Forms/X11Structs.cs
@@ -390,7 +390,7 @@ namespace System.Windows.Forms {
 		internal bool		send_event;
 		internal IntPtr		display;
 		internal IntPtr		window;
-		internal int		atom;
+		internal IntPtr		atom;
 		internal IntPtr		time;
 		internal int		state;
 	}
@@ -402,7 +402,7 @@ namespace System.Windows.Forms {
 		internal bool		send_event;
 		internal IntPtr		display;
 		internal IntPtr		window;
-		internal int		selection;
+		internal IntPtr		selection;
 		internal IntPtr		time;
 	}
 
@@ -414,9 +414,9 @@ namespace System.Windows.Forms {
 		internal IntPtr		display;
 		internal IntPtr		owner;
 		internal IntPtr		requestor;
-		internal int		selection;
-		internal int		target;
-		internal int		property;
+		internal IntPtr		selection;
+		internal IntPtr		target;
+		internal IntPtr		property;
 		internal IntPtr		time;
 	}
 
@@ -427,9 +427,9 @@ namespace System.Windows.Forms {
 		internal bool		send_event;
 		internal IntPtr		display;
 		internal IntPtr		requestor;
-		internal int		selection;
-		internal int		target;
-		internal int		property;
+		internal IntPtr		selection;
+		internal IntPtr		target;
+		internal IntPtr		property;
 		internal IntPtr		time;
 	}
 
@@ -1434,7 +1434,7 @@ namespace System.Windows.Forms {
 		internal int	X;				// Last MouseMove X coordinate; used to generate WM_MOUSEHOVER
 		internal int	Y;				// Last MouseMove Y coordinate; used to generate WM_MOUSEHOVER
 		internal int	Interval;			// in milliseconds, how long to hold before hover is generated
-		internal int	Atom;				// X Atom
+		internal IntPtr	Atom;				// X Atom
 	}
 
 	internal struct ClickStruct {
